Removed unified text for the Renew and Rebind.
I also went over the text that updates the RFC3315 and RFC3633 with
respect to Renew and Rebind processing and made a couple of changes:
- Delegating router may receive hints about the prefix length
- Client MUST send empty IAs or IAs with zeroed addresses to request the
assignment of new bindings.
- Renew and Rebind changes are in the same section because updates to
RFC3633 are common for Renew and Rebind.

This text still needs an update to address the concept of NoAddrsAvail and
NoPrefixAvail that can be returned in response to Renew and Rebind.